+++ packages/moengage_personalize/moengage_personalize_platform_interface/lib/moengage_personalize_platform_interface.dart
@@ -1,4 +1,5 @@
 export 'src/internal/constants.dart';
 export 'src/internal/payload_mapper.dart';
+export 'src/method_channel_moengage_personalize.dart';
 export 'src/model/models.dart';
 export 'src/moengage_personalize_platform_interface.dart';